[refactor] simplify modules
