+++ gradio_web/controllers/mutimodal_controllers.py
@@ -2,7 +2,7 @@ import time
 import gradio as gr
 from PIL import Image
 from modules.utils.img_segment import auto_black_by_keywords
-from modules.utils.commands import combine_commands
+from modules.utils.instruction_processing import combine_commands
 from controllers.pics_controller import change_pic_process, change_pic
 from controllers.chat_controllers import commands
 from controllers.utils_controller import submit_mask_process
