+++ gradio_web/modules/utils/image_segment.py
@@ -2,33 +2,12 @@ import base64
 from io import BytesIO
 import io
 import numpy as np
-from PIL import Image, ImageFilter
+from PIL import Image
 import gradio as gr
 
 from modules.api.pics_api import post_hgface_img_segment
 from modules.utils.image_io import trans_image_to_str, trans_str_to_image
 from const import MASK_ERODE_RATE
-
-def erode_image(image: Image.Image, erode_range: int):
-    # 将图像转换为灰度图
-    gray_image = image.convert('L')
-
-    # 使用滤波器进行腐蚀操作
-    gray_image = gray_image.filter(ImageFilter.MaxFilter(3))
-    eroded_image = gray_image.filter(ImageFilter.MinFilter(erode_range))
-
-    # 将原始图像与腐蚀后的图像进行比较，将相同位置的像素设置为黑色
-    result_image = Image.new('RGB', image.size)
-    for x in range(image.width):
-        for y in range(image.height):
-            eroded_pixel = eroded_image.getpixel((x, y))
-
-            if eroded_pixel == 0:
-                result_image.putpixel((x, y), (0, 0, 0))  # 设置为黑色
-            else:
-                result_image.putpixel((x, y), image.getpixel((x, y)))
-
-    return result_image
 
 def replace_black_pixels(image: Image):
     """
