+++ lanrui_sd.sh
@@ -1,4 +1,4 @@
 cd /app/stable-diffusion-webui/
 git apply /home/user/netdisk/data/lora_api_and_cache.patch
 cd ..
-bash start.sh --api --skip-prepare-environment --no-half
+bash start.sh --api --skip-prepare-environment
